requirements.txt
# Core dependencies
duckdb>=0.7.0
pandas>=1.3.0
numpy>=1.20.0

# Optional dependencies for export functionality
pyarrow>=7.0.0  # For Parquet export support
fastparquet>=0.8.0  # Alternative Parquet engine

# Optional dependencies for configuration
pyyaml>=6.0  # For YAML config loading

# For advanced features
typing-extensions>=4.0.0  # For better type hints support

# Development dependencies
pytest>=7.0.0  # For testing
black>=22.0.0  # For code formatting

